backend/pyproject.toml
[project]
name = "ttb-label-app-backend"
version = "0.1.0"
description = "Backend for TTB Label Verification System"
requires-python = ">=3.9"

  dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "python-jose>=3.3.0",
    "numpy>=2.0.2",
    "opencv-python>=4.11.0.86",
    "ruff>=0.6.0",
  ]

 [project.optional-dependencies]
  dev = [
    "ruff>=0.6.0",
  ]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

 [tool.hatch.build.targets.wheel]
 packages = ["app"]

 [tool.ruff]
 line-length = 88
 target-version = "py39"

 [tool.ruff.lint]
  select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
  ]
  ignore = [
    "E501", # line too long, handled by black
  ]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
